refactor(*): move presentation util under common package
